Created an Embedded Video Carousel
This carousel is a sequence of Embedded YouTube videos. These videos typically will be interviews, presentations, documentary, etc. Right now the minor issue is the spacing between the left and right arrow depending on the browser dimensions can make it difficult to change the video to full screen. Similarly, the carousel indicators should be lowered, so they are not interfering the with the video